requirements.txt
phidata
openai
python-dotenv
faiss-cpu
packaging
duckduckgo-search
fastapi
uvicorn
groq
langchain
requests
beautifulsoup4
boto3
botocore
pillow
numpy
pandas
transformers
crawl4ai
langchain-community
pinecone
streamlit